Initial commit of CaBS
… a KISS content addressable blob storage system.

CaBS uses SHA256 as a hash function to store blobs as files simmilar to e.g. git's ".git/objects" storage.
Such a storage system is useful both as a platform for higher order storage implementations and as a destination for files one doesn't wish to name explicitly.
It is planned to add features such as easy synchronisation between multiple instances of a storage as well as integrity validation functionality - i.e. something like a Merkle tree.