Implemented storage digest
`digest` returns a hash of all sorted blob hashes of a storage. This is useful
to quickly compare storage instances.

Note that this doesn't perform integrity validation and as such doesn't
assure that the blob contents are identical. This functionality was
primarily implemented to provide a quick way to e.g. detect if distributed
copies of a storage have to be synchronized.