Moved storage IO into closures
The goal was to enable storage directory definition in the UI instead of in the storage layer. This required clearer separation of IO-dependent functionality to prevent leaking the storage path all over the place.

As one can see this new solution works as advertised but increases the complexity. i.e. it once again comes down to the central problem of how to best isolate IO from the rest of the program.

Validation is now performed on the whole storage tree at once instead of trough multiple accessor functions.
`get` was expanded to better cater for multiple blob requests, i.e. the exit status is set correctly.