Added basic output to `synchronize` command
`synchronize` previously performed it's work without providing any kind of information on what it was actually doing.
Such information is useful if something fails as well as an assurance that something was actually done.

This is why `synchronize` now returns the digest of each storage, the
digest of the consensus superset of hashes and the hashes of all copied blobs.

As a consequence of these additions the storage and synchronization implementation details have been modified slightly:

* introduced `storage-missing-hash-set` record type to store digest values alongside the missing hashes
* extracted digest generation into `get-digest` function
* restricted the storage interface to only require file writer and directory reader closures
** all other closures are constructed internally as required